requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
gunicorn==20.1.0
idna==3.4
pydantic==1.10.5
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.14
